tools/code-coverage/tests/mochitest/test_coverage_specialpowers.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1380659
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 123456</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1380659 **/

  SimpleTest.waitForExplicitFinish();

  (async function() {
    await SpecialPowers.requestDumpCoverageCounters();
    SimpleTest.ok(true, "Counters dumped.");

    await SpecialPowers.requestResetCoverageCounters();
    SimpleTest.ok(true, "Counters reset.");

    SimpleTest.finish();
  })();

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1380659">Mozilla Bug 1380659</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
